requirements.txt
idna==3.4
multicolorcaptcha==1.2.0
multidict==6.0.2
Pillow==8.4.0
psutil==5.9.4
py-cord @ git+https://github.com/Pycord-Development/pycord@d9af9407bc87af7aa698ee335e979479db6ce530
pyreadline3==3.4.1
python-dotenv==0.21.0
pytz==2022.6
qrcode==7.3.1
randomstuff.py==2.0.0
requests==2.28.1
soupsieve==2.3.2.post1
timedelta==2020.12.3
typing_extensions==4.4.0
urllib3==1.26.12
utils==1.0.1
yarl==1.8.1